Fix r13897.
git-svn-id: http://skia.googlecode.com/svn/trunk@13898 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/ports/SkRemotableFontMgr_win_dw.cpp
@@ -451,7 +451,8 @@ public:
                    identity);
 
         WCHAR str[16];
-        UINT32 strLen = SkUTF16_FromUnichar(character, reinterpret_cast<uint16_t*>(str));
+        UINT32 strLen = static_cast<UINT32>(
+            SkUTF16_FromUnichar(character, reinterpret_cast<uint16_t*>(str)));
         SkTScopedComPtr<IDWriteTextLayout> fallbackLayout;
         HR_GENERAL(dwFactory->CreateTextLayout(str, strLen, fallbackFormat.get(),
                                                200.0f, 200.0f,
